Finalize split between Shared and AspNet.
