+++ src/Microsoft.AspNet.OData.Shared/Builder/LinkGenerationHelpers.cs
@@ -35,7 +35,7 @@ namespace Microsoft.AspNet.OData.Builder
                 throw Error.ArgumentNull("resourceContext");
             }
 
-            if (resourceContext.Url == null)
+            if (resourceContext.InternalUrlHelper == null)
             {
                 throw Error.Argument("resourceContext", SRResources.UrlHelperNull, typeof(ResourceContext).Name);
             }
@@ -72,7 +72,7 @@ namespace Microsoft.AspNet.OData.Builder
             {
                 throw Error.ArgumentNull("resourceContext");
             }
-            if (resourceContext.Url == null)
+            if (resourceContext.InternalUrlHelper == null)
             {
                 throw Error.Argument("resourceContext", SRResources.UrlHelperNull, typeof(ResourceContext).Name);
             }
