+++ src/Microsoft.AspNet.OData.Shared/MetadataController.cs
@@ -6,21 +6,16 @@ using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using Microsoft.AspNet.OData.Builder;
-using Microsoft.AspNet.OData.Common;
-using Microsoft.AspNet.OData.Extensions;
 using Microsoft.OData;
 using Microsoft.OData.Edm;
-using Microsoft.OData.Edm.Csdl;
 
 namespace Microsoft.AspNet.OData
 {
     /// <summary>
     /// Represents a controller for generating OData servicedoc and metadata document ($metadata).
     /// </summary>
-    public class MetadataController : ODataController
+    public partial class MetadataController : ODataController
     {
-        private static readonly Version _defaultEdmxVersion = new Version(4, 0);
-
         /// <summary>
         /// Generates the OData $metadata document.
         /// </summary>
@@ -100,18 +95,6 @@ namespace Microsoft.AspNet.OData
             };
 
             return info;
-        }
-
-        private IEdmModel GetModel()
-        {
-            IEdmModel model = Request.GetModel();
-            if (model == null)
-            {
-                throw Error.InvalidOperation(SRResources.RequestMustHaveModel);
-            }
-
-            model.SetEdmxVersion(_defaultEdmxVersion);
-            return model;
         }
     }
 }
