+++ src/Microsoft.AspNet.OData/Microsoft.AspNet.OData.csproj
@@ -89,6 +89,7 @@
     <Compile Include="Formatter\Serialization\ODataSerializerContext.cs" />
     <Compile Include="Formatter\Serialization\ODataSerializerProvider.cs" />
     <Compile Include="Formatter\Serialization\ODataSerializerProviderProxy.cs" />
+    <Compile Include="MetadataController.cs" />
     <Compile Include="Query\Expressions\ExpressionBinderBase.cs" />
     <Compile Include="Batch\UnbufferedODataBatchHandler.cs" />
     <Compile Include="Batch\ODataBatchResponseItem.cs" />
